Fix page history inconsistency when showing keywords
When toggling the keywords, the use of a hash URL made the browser
consider it a new history state, but pressing back would not restore the
previous search.

I tried to use the History API to resolve this, but it ended up being
simpler just to make the link into a toggle button for the keywords
list, and have the list appear above the entries list, not replace it.

I've also added the ability to click on a keywords in a search result
and search using it, like in the keywords list.